# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Projects RBAC API
  description: |
    REST API for authentication, groups, membership, RBAC policies, and protected assets.

    Import this file into Postman (**Import → File**), Insomnia, Thunder Client, or any OpenAPI 3 tool.
  version: 1.0.0
  contact:
    name: Projects API
servers:
  - url: http://localhost:3000
    description: Local unified Next.js app (UI + API)
  - url: "{baseUrl}"
    description: Custom / cloud deployment
    variables:
      baseUrl:
        default: https://app.example.com
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          example: alice
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          format: password
          example: Password123!
...
    SurveyQuestion:
      type: object
      required: [prompt, type]
      properties:
        questionId:
          type: string
        prompt:
          type: string
        type:
          type: string
          enum: [text, multiple_choice, yes_no]
        options:
          type: array
          items:
            type: string
        required:
          type: boolean
          default: true
    CreateSurveyRequest:
      type: object
      required: [name, questions]
      properties:
        name:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            $ref: '#/components/schemas/SurveyQuestion'
    SubmitSurveyResponseRequest:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: object
            required: [questionId, value]
            properties:
              questionId:
                type: string
              value:
                description: Text, Yes/No, or multiple-choice option
...
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: List users
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User list
        '401':
          description: Unauthorized
    post:
      tags: [Users]
      summary: Create user
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
        '400':
          description: Validation error
        '401':
          description: Unauthorized
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User found
        '401':
          description: Unauthorized
        '404':
          description: Not found
    put:
      tags: [Users]
      summary: Update user
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User updated
        '401':
          description: Unauthorized
        '404':
          description: Not found
    delete:
      tags: [Users]
      summary: Delete user
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted
        '401':
          description: Unauthorized
        '404':
          description: Not found
...
  /api/surveys:
    get:
      tags: [Surveys]
      summary: List surveys
      operationId: listSurveys
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-based)
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Page size
        - in: query
          name: search
          schema:
            type: string
          description: Case-insensitive search on name and description
        - in: query
          name: q
          schema:
            type: string
          description: Alias for search
        - in: query
          name: sort
          schema:
            type: string
            enum: [name, createdAt, updatedAt, questionCount]
            default: updatedAt
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - in: query
          name: createdBy
          schema:
            type: string
          description: Filter by creator user id
      responses:
        '200':
          description: Paginated survey list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                  page:
                    type: integer
                  limit:
                    type: integer
                  total:
                    type: integer
                  totalPages:
                    type: integer
                  sort:
                    type: string
                  order:
                    type: string
                  search:
                    type: string
                  filters:
                    type: object
                    properties:
                      createdBy:
                        type: string
                        nullable: true
        '401':
          description: Unauthorized
    post:
      tags: [Surveys]
      summary: Create a survey asset with questions
      operationId: createSurvey
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSurveyRequest'
      responses:
        '201':
          description: Survey created
        '400':
          description: Validation error
        '401':
          description: Unauthorized
  /api/surveys/{id}:
    get:
      tags: [Surveys]
      summary: Get survey by ID
      operationId: getSurvey
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Survey found
        '404':
          description: Not found
    put:
      tags: [Surveys]
      summary: Update survey
      operationId: updateSurvey
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSurveyRequest'
      responses:
        '200':
          description: Survey updated
    delete:
      tags: [Surveys]
      summary: Delete survey and its responses
      operationId: deleteSurvey
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
  /api/surveys/{id}/responses:
    get:
      tags: [Surveys]
      summary: List responses and visualization summary
      operationId: listSurveyResponses
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Responses and summary
    post:
      tags: [Surveys]
      summary: Submit survey answers as a SurveyResponse asset
      operationId: submitSurveyResponse
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitSurveyResponseRequest'
      responses:
        '201':
          description: Response saved
        '400':
          description: Validation error
...